April Homepage
New April Homepage is on Template A.
Added new colors in Variables.
Added new colors and shapes to buttons to accommodate the design of the Hero and banners.
Added new classes in generic to allow for easier transition between mobile and desktop background images.
Added new classes in generic to better manipulate text in mobile width for the banner.
Added new classes in landing to better manipulate caption text.
Added new class in landing setting align-items to stretch.