Updates for June page releases
buttons.scss -
Added a float left and float right for buttons.
Added a way to center buttons for mobile if the buttons have a float.
Changed the -primary--white styling with different colors.
Changed the -primary--transparent styling with different colors.
Changed the --long styling to have a taller button.

generic.scss:
Removed the --bg-img class.
Added hp-bannerImg-1 and hp-bannerImg-2 in order to accommodate two banner background images in the homepage.
Added a hestan-bannerImg and alfresco-bannerImg to accommodate a background banner on each of those pages.
Added a topStroke class that adds a background gradient of white and light blue for the image background for the Brand pages.
Added a col-spacing class to accommodate the space where the CSS carousel for the trade page goes since the carousel is absolute.
Changed the opacity of the carousel images when they are in the background on mobile from .3 to .1 so that the caption of the front image is more readable.

In order to truncate the website at 1920px so that the page designs did not break on higher resolutions, I made the following changes to default.scss and landing.scss
default.scss:
Added a display flex, flex-direction of column, and align-items of center
landing.scss:
Added a max-width: 1920px and a width of 100%.