Theme VDS
Removed the arrow svg on the back button from the Swatch Modal and OptionDrawers.

Changed color of cart badge from green to ocean.

Added new buttons following the new VDS design.

Changed all of the colors of the buttons around the site to ocean, charcoal, transparent, or dove.

Changed the theme of the site from green to ocean. All hover links should be ocean now.

Changed all variables for colors to start with "$clr-" in order to eliminate repeats.

Updated the sticky menu for the Shipping Policy page to add a height media query.

Updated the __header--# to match the new VDS.

Added the accessories link to the Mega Nav "/accessories/"

Added the sticky link menu functionality for the Shipping Policy page.

Added columns from 1-6 to -1-10 to accommodate new VDS designs.

Added 3-5 and 2-5 columns to accommodate new VDS designs.